+++ my-app/src/components/Welcome.tsx
